# openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_ProjectResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProjectResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[ProjectResponse]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "도메인 에러 코드. 분기는 message 가 아니라 이 값으로 한다",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "다음 페이지 커서. null 이면 마지막 페이지다. 다음 요청의 `cursor` 에 넣는다",
            "title": "Next Cursor"
          }
        },
        "required": [
          "next_cursor"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "Page_ProjectResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Page[ProjectResponse]",
        "type": "object"
      },
...
      "ProjectCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_type": {
            "default": "personal",
            "enum": [
              "personal",
              "team"
            ],
            "title": "Owner Type",
            "type": "string"
          },
          "team_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
...
      "ProjectResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "manuscript_count": {
            "default": 0,
            "title": "Manuscript Count",
            "type": "integer"
          },
          "owner_type": {
            "enum": [
              "personal",
              "team"
            ],
            "title": "Owner Type",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "team_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "title",
          "description",
          "owner_type",
          "team_id",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "type": "object"
      },
      "ProjectUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ProjectUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "성공 응답은 `{data, meta}`, 실패는 `{error: {code, message, details}}` 다. 분기는 HTTP 상태가 아니라 `error.code` 로 한다 — 같은 상태 코드에 코드가 여럿이다.\n\n입력 검증 실패는 전부 **400 `INVALID_INPUT`** 이며 `details.fields` 에 위치가 온다. 인증은 `Authorization: Bearer <access_token>`. 목록은 커서 페이지네이션이다: `meta.next_cursor` 를 다음 요청의 `cursor` 로 넘기고, null 이면 끝이다.",
    "title": "Prolog (StoryForge) API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects": {
      "get": {
        "description": "내가 직접 속한 프로젝트 + 내가 속한 팀의 팀 프로젝트. 최근에 만든 것부터.",
        "operationId": "list_projects_v1_projects_get",
        "parameters": [
          {
            "description": "페이지 크기. 기본 20 · 최대 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "페이지 크기. 기본 20 · 최대 100",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "이전 응답의 `meta.next_cursor`. 첫 페이지는 비운다",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "이전 응답의 `meta.next_cursor`. 첫 페이지는 비운다",
              "title": "Cursor"
            }
          },
          {
            "description": "소유 형태로 거른다",
            "in": "query",
            "name": "owner_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "personal",
                    "team"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "소유 형태로 거른다",
              "title": "Owner Type"
            }
          },
          {
            "description": "이 팀의 프로젝트만 본다",
            "in": "query",
            "name": "team_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "이 팀의 프로젝트만 본다",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "내 프로젝트 목록",
        "tags": [
          "PRJ"
        ]
      },
      "post": {
        "description": "만든 사람이 owner 가 된다.\n\n- 개인 프로젝트: `owner_type = personal`, `team_id` 없음.\n- 팀 프로젝트: `owner_type = team` + `team_id`. 그 팀의 팀원이면 누구나 만들 수 있고,\n  팀원 전원이 editor 로 접근한다. 팀원이 아니면 `NOT_TEAM_MEMBER` 다.\n\n`owner_type` 과 `team_id` 가 맞지 않으면 `INVALID_INPUT` 이다.",
        "operationId": "create_project_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "NOT_TEAM_MEMBER": {
                    "summary": "해당 팀의 멤버가 아닙니다",
                    "value": {
                      "error": {
                        "code": "NOT_TEAM_MEMBER",
                        "details": {},
                        "message": "해당 팀의 멤버가 아닙니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `NOT_TEAM_MEMBER` — 해당 팀의 멤버가 아닙니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "TEAM_NOT_FOUND": {
                    "summary": "팀을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "TEAM_NOT_FOUND",
                        "details": {},
                        "message": "팀을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `TEAM_NOT_FOUND` — 팀을 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "프로젝트를 만든다",
        "tags": [
          "PRJ"
        ]
      }
    },
    "/v1/projects/{projectId}": {
      "delete": {
        "description": "owner 만. 원고·챕터 등 프로젝트에 딸린 데이터가 함께 지워지며 되돌릴 수 없다.",
        "operationId": "delete_project_v1_projects__projectId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Projectid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "PROJECT_NOT_FOUND": {
                    "summary": "프로젝트를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "PROJECT_NOT_FOUND",
                        "details": {},
                        "message": "프로젝트를 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `PROJECT_NOT_FOUND` — 프로젝트를 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "프로젝트를 삭제한다",
        "tags": [
          "PRJ"
        ]
      },
      "get": {
        "description": "프로젝트의 viewer 이상. 팀 프로젝트는 팀원이면 볼 수 있다.",
        "operationId": "get_project_v1_projects__projectId__get",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Projectid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "PROJECT_NOT_FOUND": {
                    "summary": "프로젝트를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "PROJECT_NOT_FOUND",
                        "details": {},
                        "message": "프로젝트를 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `PROJECT_NOT_FOUND` — 프로젝트를 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "프로젝트를 조회한다",
        "tags": [
          "PRJ"
        ]
      },
      "patch": {
        "description": "프로젝트의 editor 이상. 보낸 필드만 바뀐다.",
        "operationId": "update_project_v1_projects__projectId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Projectid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "PROJECT_NOT_FOUND": {
                    "summary": "프로젝트를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "PROJECT_NOT_FOUND",
                        "details": {},
                        "message": "프로젝트를 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `PROJECT_NOT_FOUND` — 프로젝트를 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "프로젝트 정보를 수정한다",
        "tags": [
          "PRJ"
        ]
      }
    },
...
    "/v1/projects/{projectId}/invitations/{invitationId}": {
      "delete": {
        "description": "owner 만. 대기 중(pending)인 초대만 취소할 수 있다.",
        "operationId": "revoke_invitation_v1_projects__projectId__invitations__invitationId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Projectid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitationid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "INVITATION_NOT_FOUND": {
                    "summary": "초대를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "INVITATION_NOT_FOUND",
                        "details": {},
                        "message": "초대를 찾을 수 없습니다"
                      }
                    }
                  },
                  "PROJECT_NOT_FOUND": {
                    "summary": "프로젝트를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "PROJECT_NOT_FOUND",
                        "details": {},
                        "message": "프로젝트를 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVITATION_NOT_FOUND` — 초대를 찾을 수 없습니다\n- `PROJECT_NOT_FOUND` — 프로젝트를 찾을 수 없습니다"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "INVITATION_NOT_PENDING": {
                    "summary": "이미 처리된 초대입니다",
                    "value": {
                      "error": {
                        "code": "INVITATION_NOT_PENDING",
                        "details": {},
                        "message": "이미 처리된 초대입니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVITATION_NOT_PENDING` — 이미 처리된 초대입니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "프로젝트 초대를 취소한다",
        "tags": [
          "PRJ"
        ]
      }
    },
...
  }
}